Added Float Rectangle

+++ Engine/Vector2.cpp
@@ -52,9 +52,9 @@ Vector2_Int& Vector2_Int::operator/=( int rhs )
 	return *this = *this / rhs;
 }
 
-float Vector2_Int::GetLength() const
+int Vector2_Int::GetLength() const
 {
-	return std::sqrt( int( GetLengthSq() ) );
+	return (int)std::sqrt( int( GetLengthSq() ) );
 }
 
 int Vector2_Int::GetLengthSq() const
@@ -66,7 +66,7 @@ int Vector2_Int::GetLengthSq() const
 FLOATS
 */
 
-Vector2_Float::Vector2_Float(int x_in, int y_in)
+Vector2_Float::Vector2_Float(float x_in, float y_in)
 	:
 	x(x_in),
 	y(y_in)
@@ -84,12 +84,12 @@ Vector2_Float & Vector2_Float::operator+=(const Vector2_Float & rhs)
 	return *this = *this + rhs;
 }
 
-Vector2_Float Vector2_Float::operator*(int rhs) const
+Vector2_Float Vector2_Float::operator*(float rhs) const
 {
 	return Vector2_Float(x * rhs, y * rhs);
 }
 
-Vector2_Float & Vector2_Float::operator*=(int rhs)
+Vector2_Float & Vector2_Float::operator*=(float rhs)
 {
 	return *this = *this * rhs;
 }
@@ -104,12 +104,12 @@ Vector2_Float & Vector2_Float::operator-=(const Vector2_Float & rhs)
 	return *this = *this - rhs;
 }
 
-Vector2_Float Vector2_Float::operator/(int rhs) const
+Vector2_Float Vector2_Float::operator/(float rhs) const
 {
 	return Vector2_Float(x / rhs, y / rhs);
 }
 
-Vector2_Float & Vector2_Float::operator/=(int rhs)
+Vector2_Float & Vector2_Float::operator/=(float rhs)
 {
 	return *this = *this / rhs;
 }
@@ -119,7 +119,7 @@ float Vector2_Float::GetLength() const
 	return std::sqrt(float(GetLengthSq()));
 }
 
-int Vector2_Float::GetLengthSq() const
+float Vector2_Float::GetLengthSq() const
 {
 	return x * x + y * y;
 }
